Petits fix oubliés, oups.

+++ Assets/Scripts/Tools/RatioEngine.cs
@@ -4,11 +4,11 @@ using AFKHero.Core.Tools;
 
 public class RatioEngine : Singleton<RatioEngine> {
 
-	public double GetEnemyDamage(int damageRatio, int distance){
+	public double GetEnemyDamage(float damageRatio, float distance){
 		return (int)(damageRatio * (Mathf.Pow(distance/2,1.2f)) / 10);
 	}
 
-	public double GetEnemyHealth(int healthRatio, int distance){
+	public double GetEnemyHealth(float healthRatio, float distance){
 		return (int) (healthRatio * (Mathf.Pow (distance,1.5f)/20)-2) ;
 	}
 }
